+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/migration/battery/MigrationBatteryScreen.kt
@@ -32,6 +32,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.ui.common.provider.IsBackgroundExecutionAvailableProvider
 import co.electriccoin.zcash.ui.design.component.BlankBgScaffold
 import co.electriccoin.zcash.ui.design.component.ButtonState
+import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.design.component.ZashiButton
 import co.electriccoin.zcash.ui.design.component.ZashiButtonDefaults
 import co.electriccoin.zcash.ui.design.component.ZashiSmallTopAppBar
@@ -56,7 +57,10 @@ fun MigrationBatteryScreen() {
     val state by vm.state.collectAsStateWithLifecycle()
     val context = LocalContext.current
     val isBackgroundExecutionAvailableProvider = koinInject<IsBackgroundExecutionAvailableProvider>()
-    LceRenderer(state) { s ->
+    LceRenderer(
+        state = state,
+        loading = { isLoading -> if (isLoading && state.content == null) CircularScreenProgressIndicator() },
+    ) { s ->
         BackHandler { s.onBack() }
 
         fun isUnrestricted() = isBackgroundExecutionAvailableProvider.isAvailable()