+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/migration/notification/MigrationNotificationScreen.kt
@@ -31,6 +31,7 @@ import androidx.core.content.ContextCompat
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.ui.design.component.BlankBgScaffold
 import co.electriccoin.zcash.ui.design.component.ButtonState
+import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.design.component.ZashiButton
 import co.electriccoin.zcash.ui.design.component.ZashiButtonDefaults
 import co.electriccoin.zcash.ui.design.component.ZashiSmallTopAppBar
@@ -53,7 +54,10 @@ fun MigrationNotificationScreen() {
     val vm = koinViewModel<MigrationNotificationVM>()
     val state by vm.state.collectAsStateWithLifecycle()
     val context = LocalContext.current
-    LceRenderer(state) { s ->
+    LceRenderer(
+        state = state,
+        loading = { isLoading -> if (isLoading && state.content == null) CircularScreenProgressIndicator() },
+    ) { s ->
         BackHandler { s.onBack() }
         val isAlreadyGranted =
             remember {