+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/migration/transferreview/MigrationTransferReviewScreen.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.ui.design.component.BlankBgScaffold
 import co.electriccoin.zcash.ui.design.component.ButtonState
+import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.design.component.ZashiButton
 import co.electriccoin.zcash.ui.design.component.ZashiSmallTopAppBar
 import co.electriccoin.zcash.ui.design.component.ZashiTopAppBarBackNavigation
@@ -49,7 +50,10 @@ data object MigrationTransferReviewArgs
 fun MigrationTransferReviewScreen() {
     val vm = koinViewModel<MigrationTransferReviewVM>()
     val state by vm.state.collectAsStateWithLifecycle()
-    LceRenderer(state) { s ->
+    LceRenderer(
+        state = state,
+        loading = { isLoading -> if (isLoading && state.content == null) CircularScreenProgressIndicator() },
+    ) { s ->
         BackHandler { s.onBack() }
         MigrationTransferReviewView(s)
     }